Move set_octave and make buildup.py works with new config
