+++ application/config_api/buildup.py
@@ -1,4 +1,3 @@
-from OnMachine.SetConfig.config_path import *
 
 cluster_name = "Cluster_1"  # Write your cluster_name if version >= QOP220
 qop_ip = "192.168.50.91"  # Write the QM router IP address
@@ -31,13 +30,14 @@ port_mapping = {
 
 qubit_num = 9
 
-specs = ChannelInfo(qubit_num)
+from config_component.channel_info import ChannelInfo
+spec = ChannelInfo(qubit_num)
 
 # Set QMM
-specs.update_HardwareInfo(qop_ip=qop_ip, qop_port=qop_port, cluster_name=cluster_name)
+spec.update_HardwareInfo(qop_ip=qop_ip, qop_port=qop_port, cluster_name=cluster_name)
 # Set Octave
-specs.update_octave( "octave1", ip=qop_ip, port=11250, con="con1", port_map=port_mapping, clock="External_1000MHz")
-specs.update_octave( "octave2", ip=qop_ip, port=11249, con="con2", port_map=port_mapping, clock="External_1000MHz")
+spec.update_octave( "octave1", ip=qop_ip, port=11250, con="con1", port_map=port_mapping, clock="External_1000MHz")
+spec.update_octave( "octave2", ip=qop_ip, port=11249, con="con2", port_map=port_mapping, clock="External_1000MHz")
 
 # Only for opx+
 opxp_hardware = {
@@ -65,38 +65,48 @@ opxp_hardware = {
         2: {"offset": 0, "gain_db": 0},  # Q from down-conversion
     },
 }
-
-config = Configuration()
+from config_component.configuration import Configuration
+config_obj = Configuration()
 
 # Create controller
 from config_component.controller import controller_read_dict
-config._controllers["con1"] = controller_read_dict("con1", opxp_hardware)
-config._controllers["con2"] = controller_read_dict("con2", opxp_hardware)
+config_obj._controllers["con1"] = controller_read_dict("con1", opxp_hardware)
+config_obj._controllers["con2"] = controller_read_dict("con2", opxp_hardware)
 # Create qubit
 from config_component.construct import create_qubit, create_roChannel, create_zChannel, create_xyChannel
 
 for x_wire in [("q0",("con2",7),("con2",8)), ("q1",("con1",7),("con1",8)), ("q2",("con2",1),("con2",2)), ("q3",("con2",3),("con2",4)), ("q4",("con2",7),("con2",8)), ("q5",("con1",3),("con1",4)), ("q6",("con1",7),("con1",8)), ("q7",("con2",1),("con2",2)), ("q8",("con2",3),("con2",4))]:
-    specs.update_WireInfo_for(x_wire[0],xy_I=x_wire[1],xy_Q=x_wire[2])
+    spec.update_WireInfo_for(x_wire[0],xy_I=x_wire[1],xy_Q=x_wire[2])
 
 for z_wire in [("q0",("con1",5)), ("q1",("con1",6)), ("q2",("con1",9)), ("q3",("con1",10)), ("q4",("con1",4))]:
-    specs.update_WireInfo_for(z_wire[0],z=z_wire[1])
+    spec.update_WireInfo_for(z_wire[0],z=z_wire[1])
 
 for z_wire in [("q5",("con2",6)), ("q6",("con2",9)), ("q7",("con2",10)), ("q8",("con2",5))]:
-    specs.update_WireInfo_for(z_wire[0],z=z_wire[1])
+    spec.update_WireInfo_for(z_wire[0],z=z_wire[1])
 
 for q_idx in range(qubit_num):
     q_name = f"q{q_idx}"
-    create_qubit( config,q_name,specs.get_spec_forConfig('ro'),specs.get_spec_forConfig('xy'),specs.get_spec_forConfig('wire'),specs.get_spec_forConfig('z'))
+    create_qubit( config_obj,q_name,spec.get_spec_forConfig('ro'),spec.get_spec_forConfig('xy'),spec.get_spec_forConfig('wire'),spec.get_spec_forConfig('z'))
 
-#     create_roChannel( config, f"{q_name}_ro", specs.get_spec_forConfig('ro')[q_name],specs.get_spec_forConfig('wire')[q_name] )
-#     create_xyChannel( config, f"{q_name}_xy", specs.get_spec_forConfig('xy')[q_name],specs.get_spec_forConfig('wire')[q_name] )
-#     create_zChannel( config, f"{q_name}_z", specs.get_spec_forConfig('z')[q_name],specs.get_spec_forConfig('wire')[q_name] )
+#     create_roChannel( config, f"{q_name}_ro", spec.get_spec_forConfig('ro')[q_name],spec.get_spec_forConfig('wire')[q_name] )
+#     create_xyChannel( config, f"{q_name}_xy", spec.get_spec_forConfig('xy')[q_name],spec.get_spec_forConfig('wire')[q_name] )
+#     create_zChannel( config, f"{q_name}_z", spec.get_spec_forConfig('z')[q_name],spec.get_spec_forConfig('wire')[q_name] )
+import os
+config_path = os.path.dirname(os.path.abspath(__file__))+r'/config_link.toml'
+print( config_path )
 
-specs.export_spec(spec_loca)
-config.export_config(config_loca)
+import tomlkit
+# Assuming 'config.toml' is your file
+with open(config_path, 'r') as file:
+    content = file.read()
+    link_config = tomlkit.parse(content)
+
 
 import json
-file_path = 'output.json'
+file_path = link_config["path"]["config"]
 # Open the file in write mode and use json.dump() to export the dictionary to JSON
 with open(file_path, 'w') as json_file:
-    json.dump(config.get_config(), json_file, indent=2)
+    json.dump(config_obj.get_config(), json_file, indent=2)
+
+spec.export_spec(link_config["path"]["specification"])
+config_obj.export_config(link_config["path"]["dynamic_config"])
