+++ application/config_api/config_link.toml
@@ -1,0 +1,12 @@
+
+
+
+[sample]
+  name = "ITRI_364_scalinQ"
+
+[path]
+    config = "D:\\QM_config\\test.json"
+    dynamic_config = "D:\\QM_config\\test_config.pkl"
+    specification = "D:\\QM_config\\test_spec.pkl"
+    output_root = "D:\\Data\\format_test"
+
