+++ src/config_component/channel_info.py
@@ -1,7 +1,7 @@
 from numpy import array, cos, sin, pi, arange
 from qm.octave import QmOctaveConfig
 from qm.QuantumMachinesManager import QuantumMachinesManager
-from OnMachine.SetConfig.set_octave import OctaveUnit, octave_declaration
+from QM_driver_AS.ultitly.set_octave import OctaveUnit, octave_declaration
 
 #######################
 # AUXILIARY FUNCTIONS #
